Move ImagePicker to own file
